# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/brake-pad-wear": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "List applications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "formed_from (YYYY-MM-DD)",
                        "name": "formed_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "formed_to (YYYY-MM-DD)",
                        "name": "formed_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.applicationSummaryJSON"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/brake-pad-wear/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Get application by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.applicationFlat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Update application fields (creator)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "fields",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.applicationFlat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Soft delete application (creator)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.applicationFlat": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "driving_style": {
                    "type": "string"
                },
                "driving_style_coefficient": {
                    "type": "number"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items_count": {
                    "type": "integer"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.applicationLineFlat"
                    }
                },
                "mileage": {
                    "type": "integer"
                },
                "min_remaining_km": {
                    "type": "number"
                },
                "min_remaining_percent": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handler.applicationLineFlat": {
            "type": "object",
            "properties": {
                "base_resource_km": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "driving_style_coefficient": {
                    "type": "number"
                },
                "effective_resource_km": {
                    "type": "number"
                },
                "mileage_km": {
                    "type": "integer"
                },
                "pad_type": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "remaining_km": {
                    "type": "number"
                },
                "remaining_percent": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "service_title": {
                    "type": "string"
                }
            }
        },
        "handler.applicationSummaryJSON": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "driving_style": {
                    "type": "string"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items_count": {
                    "type": "integer"
                },
                "mileage": {
                    "type": "integer"
                },
                "min_remaining_km": {
                    "type": "number"
                },
                "min_remaining_percent": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.updateApplicationRequest": {
            "type": "object",
            "properties": {
                "driving_style": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}